Рефакторинг main.js #54
- перемещена логика работы команд бота в файл commandList.js
- исправлено удаление записей. Теперь удаляется не только папка
но и архив этой папки
