--- src/records/karpov/укрке/2023-09-23 12:41:33.txt
@@ -1,1 +1,0 @@
-кункен еунке
